Add files via upload
Data Mining, also popularly known as Knowledge Discovery in Databases (KDD), refers to the nontrivial extraction of implicit, previously unknown and potentially useful information from data in databases. 

MEAN:-
The sample mean is the average and is computed as the sum of all the observed outcomes from the sample divided by the total number of events.  We use x as the symbol for the sample mean.

MEDIAN:-
The median is the middle score.  If we have an even number of events we take the average of the two middles.  The median is better for describing the typical value.  It is often used for income and home prices.

MODE:-
The mode of a set of data is the number with the highest frequency.